\name{gcuR-package}
\alias{friendlyR-package}
\alias{friendlyR}
\docType{package}
\title{
Alternate versions of elementary statistical functions
}
\description{
This project provides alternative versions of certain elementary R functions as well as some newer functions all of which are intended to ease initial student use of R.
}
\details{
\tabular{ll}{
Package: \tab gcuR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-04\cr
License: \tab LGPL?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Joel Pitt
Maintainer: pittj@georgian.edu>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
